+++ PDFLibTests/Properties/AssemblyInfo.cs
@@ -5,31 +5,31 @@ using System.Runtime.InteropServices;
 // 組件的一般資訊是由下列的屬性集控制。
 // 變更這些屬性的值即可修改組件的相關
 // 資訊。
-[assembly: AssemblyTitle("DocWatermarkPDF")]
+[assembly: AssemblyTitle("PDFLibTests")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("DocWatermarkPDF")]
+[assembly: AssemblyProduct("PDFLibTests")]
 [assembly: AssemblyCopyright("Copyright ©  2021")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
-// 將 ComVisible 設為 false 可對 COM 元件隱藏
-// 組件中的類型。若必須從 COM 存取此組件中的類型，
+// 將 ComVisible 設定為 false 會使得這個組件中的類型 
+// 對 COM 元件而言為不可見。如果您需要從 COM 存取這個組件中 
 // 的類型，請在該類型上將 ComVisible 屬性設定為 true。
 [assembly: ComVisible(false)]
 
 // 下列 GUID 為專案公開 (Expose) 至 COM 時所要使用的 typelib ID
-[assembly: Guid("d2eb325b-747f-4de5-b249-155f67a42cdf")]
+[assembly: Guid("4ef63213-7afc-46c0-b442-1e8a737908b5")]
 
-// 組件的版本資訊由下列四個值所組成:
+// 組件的版本資訊由下列四個值所組成: 
 //
 //      主要版本
-//      次要版本
+//      次要版本 
 //      組建編號
-//      修訂
+//      修訂編號
 //
-// 您可以指定所有的值，也可以使用 '*' 將組建和修訂編號
+// 您可以指定所有的值，也可以使用 '*' 將組建和修訂編號 
 // 設為預設，如下所示:
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("1.0.0.0")]