+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ## 1.0.0 
-DocWatermarkDLL nuget 
+PDFLib nuget